backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings

python-jose
passlib[bcrypt]

motor
pymongo

librosa
numpy
torch

fpdf
reportlab

requests
python-dotenv

ipfshttpclient

openai
fpdf
requests

twilio
